requirements.txt
certifi
charset-normalizer
colorama
customtkinter
darkdetect
diffusers
filelock
huggingface-hub
idna
importlib-metadata
numpy
packaging
Pillow
pyparsing
PyYAML
regex
requests
tk
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typing_extensions
urllib3
zipp